+++ src/runtime/runtime_fe310.go
@@ -81,27 +81,4 @@ func putchar(c byte) {
 	machine.UART0.WriteByte(c)
 }
 
-func ticks() timeUnit {
-	// Combining the low bits and the high bits yields a time span of over 270
-	// years without counter rollover.
-	highBits := sifive.RTC.RTCHI.Get()
-	for {
-		lowBits := sifive.RTC.RTCLO.Get()
-		newHighBits := sifive.RTC.RTCHI.Get()
-		if newHighBits == highBits {
-			// High bits stayed the same.
-			return timeUnit(lowBits) | (timeUnit(highBits) << 32)
-		}
-		// Retry, because there was a rollover in the low bits (happening every
-		// 1.5 days).
-		highBits = newHighBits
-	}
-}
-
 const asyncScheduler = false
-
-func sleepTicks(d timeUnit) {
-	target := ticks() + d
-	for ticks() < target {
-	}
-}
